requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0